Split components

+++ app/assets/javascripts/components/Bot.es6.jsx
@@ -1,6 +1,6 @@
 class Bot extends React.Component {
-  renderButton(editing, id) {
-    if(editing) {
+  renderButton(editing, user_id, parent_id, id) {
+    if(editing && user_id && !parent_id) {
       return(
         <span className="item--button" onClick={()=>this.props.deleteBot(this.props.bot.id)}><i className="fas fa-times"></i></span>
       );
@@ -12,7 +12,7 @@ class Bot extends React.Component {
       <div className="row bots--item">
         <div className="col-12 col-sm-3 mb-2 mb-sm-0">
           <a href={`/bots/${this.props.bot.id}`}>{this.props.bot.message}</a>
-          {this.renderButton(this.props.editing, this.props.bot.id)}
+          {this.renderButton(this.props.editing, this.props.bot.user_id, this.props.bot.parent_id, this.props.bot.id)}
         </div>
         <div className="col-12 col-sm-9">
           <CodeMirrorReadOnly value={this.props.bot.response} heightAuto={true} />
